Update WD_MAP_Excel.bas
Adjusting the code to open a new Excel MAP Template that does not have code behind it that I found at Chandoo.org.  This template file works with ListObject tables.  The array of headers built in Word, CopyHeadingToArray subroutine, is transposed into the beginning of the DataBodyRange by simply referring to that cell and to the size of the range being transposed.  When inserting information into a table in this manner any current cell formula are automatically copied into to subsequent cells being inserted into the table.